# Hot path micro-optimization: NetworkTransform's HasMovedRotatedScaled