@scscgit 2020-05-02
Do you mean the performance impact, and will we use some additional metric to test it? Otherwise I thought that those 12 lines of code would be easy enough to review together. + I see that Codacy review complains about the commented out line of code - I assumed that mentioning the equivalent functionality may further improve readability, so feel free to suggest any alternative, or I'll just remove it.