@James-Frowen 2020-05-02
It seems like you have done 3 changes here
1) DistanceGreaterThan
2) Transform targetComponentTransform = targetComponent.transform;
3) local functions for moved/etc instead of calculating them all

It is hard to know from this PR how much each of these effects performance, or if any of these make it worse.