@scscgit 2020-05-02
Ok, I'll begin with the DistanceGreaterThan, which should provide at least 3% gain and possibly way more if CPU lacks sqrt instruction.

When this gets integrated, I'll commit switching to the targetComponentTransform, which should cause the most visible gain. Transform calls are included in the moved, scaled, and rotated functions, so testing them first before making the functions lazy will prevent spoiling the third commit's performance test. As the third and last commit, I'll add the local lazy functions, which will cause a gain for moving or rotating objects.