@James-Frowen 2020-05-03
I think we will try to include these 3 changes in the Experimental NetworkTransform https://github.com/vis2k/Mirror/pull/1810 instead of adding them to master right away.